+++ data/src/main/java/pjatk/s24271/jaz301/api/data/objects/MatchDAO.java
@@ -9,7 +9,7 @@ import java.math.BigInteger;
 public class MatchDAO {
     @Id
     @GeneratedValue(strategy= GenerationType.IDENTITY)
-    public int idd;
+    public Long idd;
     public String puuid;
     public String region;
     public String id;
